Move heart bitmap a few pixels up
The ATTiny memory is only 512 bytes so the display buffer canot be
stored in memory. I did not find any command for reading back the pixel
data from SSD1306 so the display has to be programmed in horizontal
bands of 8 pixels. Move the heart bitmap up a bit so it does not
overwrite text in lower band when displayed.